cambios visuales Home.jsx
