+++ client/src/pages/Registro.jsx
